requirements.txt
requests>=2.25.0
pandas>=1.3.0
openpyxl>=3.0.7
xlsxwriter>=1.4.0
python-dotenv>=0.19.0
pydantic>=1.8.2
pytest>=6.2.5
mypy>=0.910
flake8>=3.9.0
black>=21.5b2